+++ common/src/main/java/io/github/flemmli97/simplequests/quest/types/QuestBase.java
@@ -31,7 +31,6 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.function.BiFunction;
 import java.util.stream.Collectors;
 
 public abstract class QuestBase implements Comparable<QuestBase> {