+++ fabric/src/main/java/io/github/flemmli97/simplequests/fabric/SimpleQuestsFabric.java
@@ -10,6 +10,7 @@ import io.github.flemmli97.simplequests.datapack.QuestEntryRegistry;
 import io.github.flemmli97.simplequests.datapack.QuestsManager;
 import io.github.flemmli97.simplequests.network.C2SNotify;
 import io.github.flemmli97.simplequests.player.PlayerData;
+import io.github.flemmli97.simplequests.quest.QuestNumberProvider;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback;
 import net.fabricmc.fabric.api.event.Event;
@@ -22,6 +23,8 @@ import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
 import net.fabricmc.fabric.api.resource.IdentifiableResourceReloadListener;
 import net.fabricmc.fabric.api.resource.ResourceManagerHelper;
 import net.fabricmc.loader.api.FabricLoader;
+import net.minecraft.core.Registry;
+import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.server.packs.PackType;
@@ -30,11 +33,14 @@ import net.minecraft.util.profiling.ProfilerFiller;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.level.storage.loot.providers.number.LootNumberProviderType;
 
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.Executor;
 
 public class SimpleQuestsFabric implements ModInitializer {
+
+    public static LootNumberProviderType CONTEXT_MULTIPLIER;
 
     public static final Event<SimpleQuestAPI.OnQuestComplete> QUEST_COMPLETE = EventFactory.createArrayBacked(SimpleQuestAPI.OnQuestComplete.class,
             listener -> (serverPlayer, trigger, quest, progress) -> {
@@ -83,6 +89,7 @@ public class SimpleQuestsFabric implements ModInitializer {
         }));
         PayloadTypeRegistry.playC2S().register(C2SNotify.TYPE, C2SNotify.STREAM_CODEC);
         ServerPlayNetworking.registerGlobalReceiver(C2SNotify.TYPE, (pkt, ctx) -> C2SNotify.handle(pkt, ctx.player()));
+        CONTEXT_MULTIPLIER = Registry.register(BuiltInRegistries.LOOT_NUMBER_PROVIDER_TYPE, new ResourceLocation(SimpleQuests.MODID, "context_multiplier"), new LootNumberProviderType(QuestNumberProvider.ContextMultiplierNumberProvider.CODEC));
     }
 
     public static void onDeath(LivingEntity entity, DamageSource source) {