+++ neoforge/src/main/java/io/github/flemmli97/simplequests/forge/data/LangGen.java
@@ -2,8 +2,6 @@ package io.github.flemmli97.simplequests.forge.data;
 
 import com.google.common.hash.Hashing;
 import com.google.common.hash.HashingOutputStream;
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.stream.JsonWriter;